Add syncing cron
