+++ Loom/controllers/loom.js
@@ -1,24 +1,6 @@
-const config = require('../config');
 const mongoose = require('mongoose');
 const ServiceInfoModel = require('../model/serviceinfo');
-const { isOutOfSyncing, reducedErrorMessage } = require('../modules/utils');
-const { Client, CryptoUtils, LoomProvider, LocalAddress } = require('loom-js');
-const Web3 = require('web3');
-
-const privateKey = CryptoUtils.generatePrivateKey();
-const publicKey = CryptoUtils.publicKeyFromPrivateKey(privateKey);
-
-// Create the client
-const client = new Client(
-  config.client.chainId,
-  config.client.writeUrl,
-  config.client.readUrl
-);
-
-const from = LocalAddress.fromPublicKey(publicKey).toString(); // eslint-disable-line
-const loomProvider = new LoomProvider(client, privateKey);
-// Instantiate web3 client using LoomProvider as provider
-const web3 = new Web3(new LoomProvider(client, privateKey));
+const { isOutOfSyncing, reducedErrorMessage, web3, loomProvider } = require('../modules/utils');
 
 // apis
